Automatically backed up by Learn

+++ lib/cash_register.rb
@@ -26,7 +26,7 @@ class CashRegister
     if self.discount != 0
       perc_disc =((100-self.discount)/100)
     #  binding.pry
-      self.total = self.total * perc_disc.to_f
+      self.total = self.total * perc_disc.to_i
     #  binding.pry
       "After the discount, the total comes to $#{self.total}."
     else
